ctags: add the support for ctags plugin
Single tags file for a source tree:
https://vim.fandom.com/wiki/Single_tags_file_for_a_source_tree

At the top of a source tree, create a tags file by:
$ ctags -R

set the tags option in vimrc as:
    set tags=tags;
The last semicolon is the key here. When Vim tries to locate the 'tags' file, it
first looks at the current directory, then the parent directory, then the parent
of the parent, and so on. This setting works nicely with 'set autochdir',
because then Vim's current directory is the same as the directory of the file.
If you do not like using 'autochdir' but want the same upward search, use:
    set tags=./tags;
Here, the leading "./" tells Vim to use the directory of the current file rather
than Vim's working directory.